+++ src/models/app.model.ts
@@ -1,0 +1,14 @@
+export interface AppButton {
+  id: string;
+  label: string;
+  onButtonPress: () => void;
+}
+
+export enum Day {
+  Sun = 'Sunday',
+  Mon = 'Monday',
+  Tue = 'Tuesday',
+  Wed = 'Wednesday',
+  Thu = 'Thursday',
+  Fri = 'Friday',
+}
