+++ src/models/meals.model.ts
@@ -6,18 +6,3 @@ export enum MealCategories {
 
 export type MealText = string;
 export type Category = string;
-
-export interface AppButton {
-  id: string;
-  label: string;
-  onButtonPress: () => void;
-}
-
-export enum Day {
-  Sun = 'Sunday',
-  Mon = 'Monday',
-  Tue = 'Tuesday',
-  Wed = 'Wednesday',
-  Thu = 'Thursday',
-  Fri = 'Friday',
-}
